Various fixes and improvements
- Reworked weapons (reduced Toaster ammo consumption)
- Added Bird fire sound
- Fixed Gem Ring
- Fixed Pole
- Fixed Bridge
- Reworked collisions
- Reworked sounds
- Refactoring and removed unused code

+++ Jazz2.Core/Game/Structs/LevelInitialization.cs
@@ -26,7 +26,7 @@ namespace Jazz2.Game.Structs
         public PlayerType Type;
 
         public int Lives;
-        public int[] Ammo;
+        public short[] Ammo;
         public byte[] WeaponUpgrades;
         public int Score;
         public int FoodEaten;
